requirements.txt
Twisted==16.6.0
requests==2.20.1
Flask==1.0.2
google-cloud-storage==1.13.0
tf-nightly-2.0-preview>=2.0.0-dev20190427